Rename and add test
