+++ direct/data/mri_transforms.py
@@ -10,7 +10,7 @@ from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple, Union
 import numpy as np
 import torch
 
-from direct.algorithms.mri import EspiritCalibration
+from direct.algorithms.mri_algorithms import EspiritCalibration
 from direct.data import transforms as T
 from direct.exceptions import ItemNotFoundException
 from direct.types import DirectEnum, KspaceKey
